+++ packages/moonstone/CHANGELOG.md
@@ -25,7 +25,7 @@ The following is a curated list of changes in the Enact moonstone module, newest
 - `moonstone/Slider` by removing unnecessary repaints to the screen
 - `moonstone/VideoPlayer` to correctly position knob when interacting with media slider
 - `moonstone/VideoPlayer` to not read out the focused button when the media controls hide
-- `moonstone/MarqueeDecorator` to stop when unhovering a disabled component using `marqueeOn` `'focus'`
+- `moonstone/MarqueeDecorator` to correctly stop or continue when a focused `marqueeOn` `'focus'` component becomes disabled
 
 ## [1.12.2] - 2017-11-15